refactor: make identities common
