+++ src/Sinch/Conversation/Common/ChannelIdentity.cs
@@ -1,17 +1,5 @@
-using System.Collections.Generic;
-
-namespace Sinch.Conversation.Messages
+﻿namespace Sinch.Conversation.Common
 {
-    public class Identified : IRecipient
-    {
-        public IdentifiedBy IdentifiedBy { get; set; }
-    }
-
-    public class IdentifiedBy
-    {
-        public List<ChannelIdentity> ChannelIdentities { get; set; }
-    }
-
     public class ChannelIdentity
     {
         /// <summary>
