+++ src/Sinch/Conversation/Contacts/GetChannelProfile/GetChannelProfileRequest.cs
@@ -1,4 +1,5 @@
-﻿using Sinch.Conversation.Messages;
+﻿using Sinch.Conversation.Common;
+using Sinch.Conversation.Messages;
 
 namespace Sinch.Conversation.Contacts.GetChannelProfile
 {
