+++ src/Sinch/Conversation/Conversations/InjectMessage/InjectMessageRequest.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Text;
 using System.Text.Json.Serialization;
+using Sinch.Conversation.Common;
 using Sinch.Conversation.Messages;
 using Sinch.Conversation.Messages.Message;
