+++ tests/Sinch.Tests/Conversation/ContactsTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using FluentAssertions;
 using Sinch.Conversation;
+using Sinch.Conversation.Common;
 using Sinch.Conversation.Contacts;
 using Sinch.Conversation.Messages;
 using Xunit;
