+++ tests/Sinch.Tests/e2e/Conversation/ContactsTests.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using FluentAssertions;
 using Sinch.Conversation;
+using Sinch.Conversation.Common;
 using Sinch.Conversation.Contacts.Create;
 using Sinch.Conversation.Contacts.GetChannelProfile;
 using Sinch.Conversation.Contacts.List;
